LPS-56789 Move resources
