+++ modules/apps/portlet-display-template/portlet-display-template-service/src/com/liferay/portlet/display/template/service/PortletDisplayTemplateConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.portletdisplaytemplate.util;
+package com.liferay.portlet.display.template.service;
 
 import com.liferay.portal.kernel.template.TemplateConstants;
