+++ modules/apps/portlet-display-template/portlet-display-template-service/src/com/liferay/portlet/display/template/service/impl/PortletDisplayTemplateImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.portletdisplaytemplate.util;
+package com.liferay.portlet.display.template.service.impl;
 
 import com.liferay.portal.kernel.bean.ClassLoaderBeanHandler;
 import com.liferay.portal.kernel.exception.PortalException;
@@ -39,6 +39,8 @@ import com.liferay.portal.templateparser.Transformer;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PortletKeys;
 import com.liferay.portlet.PortletURLUtil;
+import com.liferay.portlet.display.template.service.PortletDisplayTemplate;
+import com.liferay.portlet.display.template.service.PortletDisplayTemplateConstants;
 import com.liferay.portlet.dynamicdatamapping.NoSuchTemplateException;
 import com.liferay.portlet.dynamicdatamapping.model.DDMTemplate;
 import com.liferay.portlet.dynamicdatamapping.service.DDMTemplateLocalServiceUtil;
@@ -46,7 +48,6 @@ import com.liferay.taglib.servlet.PipingServletResponse;
 import com.liferay.taglib.util.VelocityTaglib;
 
 import java.lang.reflect.InvocationHandler;
-
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
@@ -58,17 +59,20 @@ import javax.portlet.PortletPreferences;
 import javax.portlet.PortletURL;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
-
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import org.osgi.service.component.annotations.Component;
 
 /**
  * @author Eduardo Garcia
  * @author Juan Fernández
  * @author Brian Wing Shun Chan
  * @author Raymond Augé
+ * @author Leonardo Barros
  */
 @DoPrivileged
+@Component(immediate = true)
 public class PortletDisplayTemplateImpl implements PortletDisplayTemplate {
 
 	@Override
