--- portal-service/src/com/liferay/portlet/portletdisplaytemplate/util/packageinfo
@@ -1,1 +1,0 @@
-version 7.0.0
